==> model/06-validation.ipynb <==
# IGP 5 Models - Validate Final Models



#!pip install --upgrade scikit-learn==1.4.2

# load functions in python file with magic command
%run ../code/final-model.py

output_csv_path = '../output/'

# read validation from csv
male_X_valid_scaled = pd.read_csv(output_csv_path + 'male_X_valid_scaled.csv')
male_y_valid = pd.read_csv(output_csv_path + 'male_y_valid.csv').iloc[:, 0]
female_X_valid_scaled = pd.read_csv(output_csv_path + 'female_X_valid_scaled.csv')
female_y_valid = pd.read_csv(output_csv_path + 'female_y_valid.csv').iloc[:, 0]
both_X_valid_scaled = pd.read_csv(output_csv_path + 'both_X_valid_scaled.csv')
both_y_valid = pd.read_csv(output_csv_path + 'both_y_valid.csv').iloc[:, 0]

# drop features to match models
male_X_valid_scaled_2 = male_X_valid_scaled.iloc[:, [2, 3]]
male_X_valid_scaled_3 = male_X_valid_scaled.iloc[:, [2, 3,9]]
female_X_valid_scaled_2 = female_X_valid_scaled.iloc[:, [2, 3]]
female_X_valid_scaled_3 = female_X_valid_scaled.iloc[:, [2, 3, 5]]
both_X_valid_scaled_2 = both_X_valid_scaled.iloc[:, [2, 3]]
both_X_valid_scaled_3 = both_X_valid_scaled.iloc[:, [2, 3, 6]]

# load models
male_2_model = joblib.load('male_2_feature_lgbm_model.pkl')
male_3_model = joblib.load('male_3_feature_svc_model.pkl')
female_2_model = joblib.load('female_2_feature_gradboost_model.pkl')
female_3_model = joblib.load('female_3_feature_gradboost_model.pkl')
both_2_model = joblib.load('both_2_feature_gradboost_model.pkl')
both_3_model = joblib.load('both_3_feature_rf_model.pkl')


# models and validation data
models = [
    ('male_2', male_2_model, male_X_valid_scaled_2, male_y_valid),
    ('male_3', male_3_model, male_X_valid_scaled_3, male_y_valid),
    ('female_2', female_2_model, female_X_valid_scaled_2, female_y_valid),
    ('female_3', female_3_model, female_X_valid_scaled_3, female_y_valid),
    ('both_2', both_2_model, both_X_valid_scaled_2, both_y_valid),
    ('both_3', both_3_model, both_X_valid_scaled_3, both_y_valid)
]

#from sklearn.metrics import accuracy_score, f1_score, matthews_corrcoef



# predictions and  metrics
for model_name, model, X_valid, y_valid in models:
    preds = model.predict(X_valid)
    accuracy = accuracy_score(y_valid, preds)
    f1 = f1_score(y_valid, preds)
    mcc = matthews_corrcoef(y_valid, preds)
    
    print(f"{model_name} Model:")
    print(f"Accuracy: {accuracy:.4f}")
    print(f"F1 Score: {f1:.4f}")
    print(f"MCC: {mcc:.4f}")
    print()
